requirements.txt
langchain>=0.1.0
langchain-community>=0.0.16
langchain-core>=0.3.39
langchain-openai>=0.3.7
tiktoken>=0.7.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
InstructorEmbedding
streamlit>=1.32.0
python-dotenv
numpy<2.0.0
torch>=2.1.0
PyPDF2
chromadb
transformers
huggingface_hub==0.21.4
torchvision==0.16.0
nest-asyncio==1.5.8
altair==4
google-genai
google-generativeai